requirements.txt
amqp==2.4.2
autopep8==1.4.3
babel==2.6.0
billiard==3.5.0.5
celery==4.2.1
certifi==2019.3.9
chardet==3.0.4
django==2.1.7
djangorestframework==3.9.2
entrypoints==0.3
flake8==3.7.7
flower==0.9.3
idna==2.8
kombu==4.4.0
mccabe==0.6.1
oauthlib==3.0.1
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2018.9
requests-oauthlib==1.2.0
requests==2.21.0
tornado==5.1.1
urllib3==1.24.1
vine==1.3.0
